fix loss and labels for multiclass

+++ src/custom_trainer.py
@@ -27,14 +27,12 @@ class MultiOutputTrainer(Trainer):
         logits = outputs.get("logits")
 
         loss_fct = nn.CrossEntropyLoss().to(model.device)
-        
-        #print(logits[:, 0:3], '\n', labels[:, 0])
 
-        loss_1 = loss_fct(logits[:, 0:3], labels[:, 0].long())
-        loss_2 = loss_fct(logits[:, 3:6], labels[:, 1].long())
-        loss_3 = loss_fct(logits[:, 6:9], labels[:, 2].long())
-        loss_4 = loss_fct(logits[:, 9:12], labels[:, 3].long())
-        loss_5 = loss_fct(logits[:, 12:], labels[:, 4].long())
+        loss_1 = loss_fct(logits[:, 0:3], labels[:, 0])
+        loss_2 = loss_fct(logits[:, 3:6], labels[:, 1])
+        loss_3 = loss_fct(logits[:, 6:9], labels[:, 2])
+        loss_4 = loss_fct(logits[:, 9:12], labels[:, 3])
+        loss_5 = loss_fct(logits[:, 12:], labels[:, 4])
 
         loss = (
             loss_1 + loss_2 + loss_3 + loss_4 + loss_5
